Update tests for Relays
