+++ Tests/RxCocoaTests/Observable+BindTests.swift
@@ -66,86 +66,6 @@ extension ObservableBindTest {
     }
 }
 
-// MARK: bind(to:) publish relay
-
-extension ObservableBindTest {
-    func testBindToPublishRelay() {
-        var events: [Recorded<Event<Int>>] = []
-        
-        let relay = PublishRelay<Int>()
-        
-        _ = relay.subscribe{ event in
-            events.append(Recorded(time: 0, value: event))
-        }
-        
-        _ = Observable.just(1).bind(to: relay)
-        
-        XCTAssertEqual(events, [
-            .next(1)
-            ])
-    }
-    
-    func testBindToOptionalPublishRelay() {
-        var events: [Recorded<Event<Int?>>] = []
-        
-        let relay = PublishRelay<Int?>()
-        
-        _ = relay.subscribe{ event in
-            events.append(Recorded(time: 0, value: event))
-        }
-
-        _ = (Observable.just(1) as Observable<Int>).bind(to: relay)
-
-        XCTAssertEqual(events, [
-            .next(1)
-            ])
-    }
-
-    func testBindToPublishRelayNoAmbiguity() {
-        var events: [Recorded<Event<Int?>>] = []
-        
-        let relay = PublishRelay<Int?>()
-        
-        _ = relay.subscribe{ event in
-            events.append(Recorded(time: 0, value: event))
-        }
-
-        _ = Observable.just(1).bind(to: relay)
-
-        XCTAssertEqual(events, [
-            .next(1)
-            ])
-    }
-}
-
-// MARK: bind(to:) behavior relay
-
-extension ObservableBindTest {
-    func testBindToBehaviorRelay() {
-        let relay = BehaviorRelay<Int>(value: 0)
-        
-        _ = Observable.just(1).bind(to: relay)
-        
-        XCTAssertEqual(relay.value, 1)
-    }
-    
-    func testBindToOptionalBehaviorRelay() {
-        let relay = BehaviorRelay<Int?>(value: 0)
-
-        _ = (Observable.just(1) as Observable<Int>).bind(to: relay)
-
-        XCTAssertEqual(relay.value, 1)
-    }
-
-    func testBindToBehaviorRelayNoAmbiguity() {
-        let relay = BehaviorRelay<Int?>(value: 0)
-
-        _ = Observable.just(1).bind(to: relay)
-
-        XCTAssertEqual(relay.value, 1)
-    }
-}
-
 // MARK: bind(to:) curried
 
 extension ObservableBindTest {
