+++ Tests/RxCocoaTests/RxTest+Controls.swift
@@ -8,6 +8,7 @@
 
 import RxCocoa
 import RxSwift
+import RxRelay
 import XCTest
 
 extension RxTest {
